auth-service/requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
python-jose==3.5.0
passlib==1.7.4
psycopg[binary]==3.2.6
SQLAlchemy==2.0.46
python-multipart==0.0.22
python-dotenv==1.0.0